requirements.txt
streamlit>=1.30
pyserial>=3.5
numpy>=1.24
scipy>=1.11
pandas>=2.0
plotly>=5.17
bleak>=0.22
kaleido>=0.2
